new package; split up heartbeat class into lifecycle manager and actual requester/handler
