+++ app/src/org/commcare/heartbeat/HeartbeatRequester.java
@@ -1,8 +1,10 @@
-package org.commcare;
+package org.commcare.heartbeat;
 
 import android.os.Handler;
 import android.os.Looper;
 
+import org.commcare.CommCareApp;
+import org.commcare.CommCareApplication;
 import org.commcare.core.interfaces.HttpResponseProcessor;
 import org.commcare.core.network.ModernHttpRequester;
 import org.commcare.logging.AndroidLogger;
@@ -22,23 +24,12 @@ import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.Date;
 import java.util.HashMap;
-import java.util.Timer;
-import java.util.TimerTask;
 
 /**
- * While active, this class is responsible for using a TimerTask to periodically ping the server
- * with a "heartbeat" request, and then handle the response. The lifecycle of the TimerTask is
- * tied to that of the CommCareSessionService; it should be started whenever a session service is
- * started, and ended whenever a session service is ended for any reason.
- *
- * Currently, the primary content of the server's response to the heartbeat request will be
- * information about potential binary or app updates that the app should prompt users to conduct.
- *
- * Created by amstone326 on 4/13/17.
+ * Created by amstone326 on 5/5/17.
  */
-public class HeartbeatLifecycleManager {
 
-    private static final long ONE_DAY_IN_MS = 24 * 60 * 60 * 1000;
+public class HeartbeatRequester {
 
     private static final String TEST_RESPONSE =
             "{\"latest_apk_version\":{\"value\":\"2.36.1\"},\"latest_ccz_version\":{\"value\":\"197\", \"force_by_date\":\"2017-04-24\"}}";
@@ -46,8 +37,6 @@ public class HeartbeatLifecycleManager {
     private static final String QUARANTINED_FORMS_PARAM = "num_quarantined_forms";
     private static final String UNSENT_FORMS_PARAM = "num_unsent_forms";
     private static final String LAST_SYNC_TIME_PARAM = "last_sync_time";
-
-    private Timer heartbeatTimer;
 
     private final HttpResponseProcessor responseProcessor = new HttpResponseProcessor() {
 
@@ -101,49 +90,7 @@ public class HeartbeatLifecycleManager {
         }
     };
 
-    private static HeartbeatLifecycleManager INSTANCE;
-    public static HeartbeatLifecycleManager instance() {
-        if (INSTANCE == null) {
-            INSTANCE = new HeartbeatLifecycleManager();
-        }
-        return INSTANCE;
-    }
-
-    public void startHeartbeatCommunications() {
-        if (heartbeatTimer != null) {
-            // Make sure we end anything still in progress
-            heartbeatTimer.cancel();
-        }
-        heartbeatTimer = new Timer();
-        heartbeatTimer.schedule(new TimerTask() {
-            @Override
-            public void run() {
-                try {
-                    User currentUser = CommCareApplication.instance().getSession().getLoggedInUser();
-                    //simulateRequestGettingStuck();
-                    //requestHeartbeat(currentUser);
-                    parseTestHeartbeatResponse();
-                } catch (SessionUnavailableException e) {
-                    // Means the session has ended, so we should stop these requests
-                    stopHeartbeatCommunications();
-                } catch (Exception e) {
-                    // Encountered a different, unexpected issue
-                    stopHeartbeatCommunications();
-                    Logger.log(AndroidLogger.TYPE_ERROR_SERVER_COMMS,
-                            "Encountered unexpected exception during heartbeat communications: "
-                                    + e.getMessage() + ". Stopping the heartbeat thread.");
-                }
-            }
-        }, new Date(), ONE_DAY_IN_MS);
-    }
-
-    public void stopHeartbeatCommunications() {
-        if (heartbeatTimer != null) {
-            heartbeatTimer.cancel();
-        }
-    }
-
-    public static void parseTestHeartbeatResponse() {
+    protected static void parseTestHeartbeatResponse() {
         System.out.println("NOTE: Testing heartbeat response processing");
         try {
             parseHeartbeatResponse(new JSONObject(TEST_RESPONSE));
@@ -152,7 +99,7 @@ public class HeartbeatLifecycleManager {
         }
     }
 
-    private static void simulateRequestGettingStuck() {
+    protected static void simulateRequestGettingStuck() {
         System.out.println("Before sleeping");
         try {
             Thread.sleep(5*1000);
@@ -162,14 +109,14 @@ public class HeartbeatLifecycleManager {
         System.out.println("After sleeping");
     }
 
-    private void requestHeartbeat(User currentUser) {
+    protected void requestHeartbeat(User currentUser) {
         CommCareApp currentApp = CommCareApplication.instance().getCurrentApp();
         String urlString = currentApp.getAppPreferences().getString(
                 CommCareServerPreferences.PREFS_HEARTBEAT_URL_KEY, null);
         if (urlString == null) {
             // This app was generated before the heartbeat URL started being included, so we
             // can't make the request
-            stopHeartbeatCommunications();
+            HeartbeatLifecycleManager.instance().stopHeartbeatCommunications();
             return;
         }
 
@@ -250,5 +197,4 @@ public class HeartbeatLifecycleManager {
                             "UpdateToPrompt object : " + e.getMessage());
         }
     }
-
 }
